+++ mobile/src/components/List/index.tsx
@@ -5,7 +5,7 @@ import { Pressable } from "react-native";
 import { cn } from "~/lib/style";
 import type { ListItemContentProps } from "./ListItemContent";
 import { ListItemContent } from "./ListItemContent";
-import type { PressProps } from "../../Form/Button";
+import type { PressProps } from "../Form/Button";
 
 export type ListItemProps = ListItemContentProps &
   PressProps & {
