+++ mobile/src/navigation/screens/settings/PlaybackSettingsView.tsx
@@ -6,8 +6,8 @@ import { PreferenceTogglers } from "~/stores/Preference/actions";
 import { StandardScrollLayout } from "~/navigation/layouts/StandardScroll";
 import { PlaybackDelaySheet } from "~/navigation/sheets/PlaybackDelaySheet";
 
-import { SegmentedList } from "~/components/DS-2/List/Segmented";
 import { Switch } from "~/components/Form/Switch";
+import { SegmentedList } from "~/components/List/Segmented";
 import { useSheetRef } from "~/components/Sheet/useSheetRef";
 
 export default function PlaybackSettings() {
