+++ mobile/src/navigation/screens/settings/View.tsx
@@ -13,9 +13,9 @@ import { LanguageSheet } from "~/navigation/sheets/LanguageSheet";
 
 import { APP_VERSION } from "~/constants/Config";
 import * as LINKS from "~/constants/Links";
-import { SegmentedList } from "~/components/DS-2/List/Segmented";
 import { Divider } from "~/components/Divider";
 import { Switch } from "~/components/Form/Switch";
+import { SegmentedList } from "~/components/List/Segmented";
 import { useSheetRef } from "~/components/Sheet/useSheetRef";
 import { LANGUAGES } from "~/modules/i18n/constants";
